Updated the Mapbox instructions
You no longer need to specify a project id, but rather a map id (ex: mapbox.streets, mapbox.satellite, etc). So I updated the tutorial to reflect this. I also updated some Mapbox URLs and made it more clear on how you attain a Mapbox API access token and the URL you go to in order to retrieve that token.